+++ src/main/java/io/github/thepieterdc/random/numerical/RandomPrimeGenerator.java
@@ -104,7 +104,7 @@ public class RandomPrimeGenerator extends AbstractRandomGenerator<Integer> {
 	}
 	
 	@Override
-	public int getCapacity() {
+	public long getCapacity() {
 		return PRIMES.length;
 	}
 }
